+++ include/ParallaxGenDirectory/ParallaxGenDirectory.hpp
@@ -15,9 +15,37 @@ private:
 	std::vector<std::filesystem::path> heightMaps;
 	std::vector<std::filesystem::path> complexMaterialMaps;
 	std::vector<std::filesystem::path> meshes;
-  std::vector<nlohmann::json> truePBRConfigs;
+  	std::vector<nlohmann::json> truePBRConfigs;
   
 	nlohmann::json PG_config;
+
+	// Validation is hardcoded to make it user-proof. Match the schema of the config, but instead
+	// of array leafs, string leafs with regex patterns for acceptable list values are used
+	static inline const std::string PG_config_validation_raw = R"(
+	{
+		"parallax_lookup": {
+			"archive_blocklist": "^[^\\/\\\\]*$",
+			"allowlist": ".*_.*\\.dds$",
+			"blocklist": ".*"
+		},
+		"complexmaterial_lookup": {
+			"archive_blocklist": "^[^\\/\\\\]*$",
+			"allowlist": ".*_.*\\.dds$",
+			"blocklist": ".*"
+		},
+		"truepbr_cfg_lookup": {
+			"archive_blocklist": "^[^\\/\\\\]*$",
+			"allowlist": ".*\\.json$",
+			"blocklist": ".*"
+		},
+		"nif_lookup": {
+			"archive_blocklist": "^[^\\/\\\\]*$",
+			"allowlist": ".*\\.nif$",
+			"blocklist": ".*"
+		}
+	}
+	)";
+	static inline const nlohmann::json PG_config_validation = nlohmann::json::parse(PG_config_validation_raw);
 
 	static inline const std::vector<std::string> truePBR_filename_fields = { "match_normal", "match_diffuse", "rename" };
 
@@ -57,8 +85,13 @@ public:
 	const std::vector<std::filesystem::path> getMeshes() const;
 	const std::vector<nlohmann::json> getTruePBRConfigs() const;
 
+	const std::string getHeightMapFromBase(const std::string& base) const;
+	const std::string getComplexMaterialMapFromBase(const std::string& base) const;
+
 	// Helpers
-	static void merge_json_smart(nlohmann::json& target, const nlohmann::json& source);
+	static void merge_json_smart(nlohmann::json& target, const nlohmann::json& source, const nlohmann::json& validation);
+	static bool validate_json(const nlohmann::json& item, const nlohmann::json& validation, const std::string& key);
 	static std::vector<std::wstring> jsonArrayToWString(const nlohmann::json& json_array);
 	static void replaceForwardSlashes(nlohmann::json& j);
+	static std::filesystem::path matchBase(const std::string& base, const std::vector<std::filesystem::path>& search_list);
 };